Move Authentication related Activities to com.github.versus.auth
