+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:theme="@style/Theme.Versus"
         tools:targetApi="31">
         <activity
-            android:name=".LogInActivity"
+            android:name=".auth.LogInActivity"
             android:exported="false" />
         <activity
-            android:name=".SignInActivity"
+            android:name=".auth.SignInActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
