+++ app/src/main/java/com/github/versus/EntryActivity.java
@@ -5,6 +5,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 
+import com.github.versus.auth.LogInActivity;
 import com.google.firebase.auth.FirebaseAuth;
 
 import static java.util.Objects.*;
