+++ app/src/main/java/com/github/versus/auth/LogInActivity.java
@@ -1,4 +1,4 @@
-package com.github.versus;
+package com.github.versus.auth;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -7,14 +7,14 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.github.versus.MainActivity;
+import com.github.versus.R;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
 public final class LogInActivity extends AppCompatActivity {
